# apps/api-core/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ClinCommand OS™ Enterprise API Specifications
  version: 1.0.0
  description: OpenAPI 3.1.0 specification for identity management, SCIM user syncing, SaaS billing, AI risk governance, and GxP validation.
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
paths:
  /auth/login:
    post:
      summary: Authenticate User
      description: Logs in a user and returns a short-lived access token along with a rotating refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin_gxp
                password:
                  type: string
                  format: password
                  example: securePassword123
      responses:
        '200':
          description: Successful authentication
          headers:
            Set-Cookie:
              schema:
                type: string
                description: HTTP-only session refresh token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT Access Token
                  refreshToken:
                    type: string
                    description: Rotated Refresh Token
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: The requested input validation failed.
        details:
          type: object